+++ Types/Emoji.Symbol/get_HTML.ps1
@@ -1,0 +1,11 @@
+<#
+.SYNOPSIS
+    Gets an Emoji as HTML 
+.DESCRIPTION
+    Gets an Emoji as an HTML span
+.EXAMPLE
+    (Get-Emoji -Name 'Grinning Face').HTML
+#>
+param()
+
+"<span>$([Web.HttpUtility]::HtmlEncode($this.Emoji))</span>"